+++ paddle/fluid/distributed/collective/process_group_gloo.h
@@ -20,13 +20,11 @@
 
 #include "paddle/fluid/distributed/collective/process_group.h"
 #include "paddle/fluid/distributed/collective/process_group_without_stream.h"
+#include "paddle/phi/core/distributed/store/store.h"
 
 #ifdef PADDLE_WITH_GLOO
 #include "paddle/fluid/framework/fleet/gloo_wrapper.h"
 #endif
-
-#include "paddle/phi/core/store/store.h"
-#include "paddle/phi/core/store/tcp_store.h"
 
 namespace paddle {
 namespace distributed {
