+++ paddle/fluid/distributed/collective/process_group_mpi.h
@@ -15,15 +15,14 @@
 #pragma once
 
 #include <chrono>
-#include <map>
-#include <string>
-#include <unordered_map>
-#include <vector>
-
 #include <condition_variable>
 #include <deque>
 #include <exception>
+#include <map>
 #include <mutex>
+#include <string>
+#include <unordered_map>
+#include <vector>
 
 #include "paddle/fluid/distributed/collective/process_group.h"
 #include "paddle/fluid/distributed/collective/process_group_without_stream.h"
