+++ paddle/phi/core/CMakeLists.txt
@@ -1,6 +1,6 @@
 # compatible utils used for fluid op system
 add_subdirectory(compat)
-add_subdirectory(store)
+add_subdirectory(distributed)
 
 if(WITH_GPU)
   proto_library(external_error_proto SRCS external_error.proto)
@@ -141,17 +141,3 @@ else()
     SRCS tensor_utils.cc
     DEPS dense_tensor selected_rows malloc memcpy device_context phi_backends)
 endif()
-
-set(COMM_CONTEXT_MANAGER_DEPS tcp_store)
-if(WITH_NCCL OR WITH_RCCL)
-  cc_library(
-    nccl_comm_context
-    SRCS nccl_comm_context.cc
-    DEPS dense_tensor)
-  list(APPEND COMM_CONTEXT_MANAGER_DEPS nccl_comm_context)
-endif()
-
-cc_library(
-  comm_context_manager
-  SRCS comm_context_manager.cc
-  DEPS ${COMM_CONTEXT_MANAGER_DEPS})
