+++ paddle/phi/core/distributed/store/store.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/phi/core/store/store.h"
+#include "paddle/phi/core/distributed/store/store.h"
 #include "paddle/phi/core/enforce.h"
 
 namespace phi {
